Initial approach
This script takes the first line of the playlist file as the path to
the next song to play.
If the file is empty it waits 1 second.

At the same time, this script can append new songs to the playlist file
while the player is running, this can be done thanks to the lock file
which works as a semaphore to control access to the playlist file.

If this script is added to the PATH, it can be accessed from
ranger to append songs while in other terminal the service is
playing or waiting for a new song.